+++ src/test/java/com/github/alexdlaird/ngrok/NgrokClientTest.java
@@ -42,11 +42,11 @@ import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 
 import static com.github.alexdlaird.util.StringUtils.isNotBlank;
-import static java.net.HttpURLConnection.HTTP_BAD_GATEWAY;
 import static org.hamcrest.CoreMatchers.startsWith;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.greaterThan;
@@ -56,7 +56,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assumptions.assumeTrue;
 
@@ -83,10 +83,10 @@ class NgrokClientTest extends NgrokTestCase {
     @Test
     public void testGetters() {
         // THEN
-        assertEquals(javaNgrokConfigV2, ngrokClientV2.getJavaNgrokConfig());
-        assertEquals(ngrokProcessV2, ngrokClientV2.getNgrokProcess());
-        assertEquals(ngrokInstaller, ngrokClientV2.getNgrokProcess().getNgrokInstaller());
-        assertNotNull(ngrokClientV2.getHttpClient());
+        assertEquals(javaNgrokConfigV3, ngrokClientV3.getJavaNgrokConfig());
+        assertEquals(ngrokProcessV3, ngrokClientV3.getNgrokProcess());
+        assertEquals(ngrokInstaller, ngrokClientV3.getNgrokProcess().getNgrokInstaller());
+        assertNotNull(ngrokClientV3.getHttpClient());
     }
 
     @Test
@@ -107,6 +107,7 @@ class NgrokClientTest extends NgrokTestCase {
         // THEN
         assertTrue(ngrokClientV2.getNgrokProcess().getVersion().startsWith("2"));
         assertTrue(ngrokClientV2.getNgrokProcess().isRunning());
+        assertNull(tunnel.getId());
         assertThat(tunnel.getName(), startsWith("http-5000-"));
         assertEquals("http", tunnel.getProto());
         assertEquals("http://localhost:5000", tunnel.getConfig().getAddr());
@@ -141,6 +142,7 @@ class NgrokClientTest extends NgrokTestCase {
         // THEN
         assertTrue(ngrokClientV3.getNgrokProcess().getVersion().startsWith("3"));
         assertTrue(ngrokClientV3.getNgrokProcess().isRunning());
+        assertNotNull(tunnel.getId());
         assertThat(tunnel.getName(), startsWith("http-5000-"));
         assertEquals("https", tunnel.getProto());
         assertEquals("http://localhost:5000", tunnel.getConfig().getAddr());
@@ -563,30 +565,30 @@ class NgrokClientTest extends NgrokTestCase {
         assertEquals("https", tunnel.getProto());
         assertEquals("file:///", tunnel.getConfig().getAddr());
         assertNotNull(tunnel.getPublicUrl());
-        assertThat(tunnel.getPublicUrl(), startsWith("http://"));
+        assertThat(tunnel.getPublicUrl(), startsWith("https://"));
     }
 
     @Test
-    public void testDisconnectFileserver() throws InterruptedException {
+    public void testDisconnectFileserverV2() throws InterruptedException {
         final String ngrokAuthToken = System.getenv("NGROK_AUTHTOKEN");
         assumeTrue(isNotBlank(System.getenv("NGROK_AUTHTOKEN")), "NGROK_AUTHTOKEN environment variable not set");
 
         // GIVEN
-        ngrokClientV3.getNgrokProcess().setAuthToken(ngrokAuthToken);
-        assertFalse(ngrokClientV3.getNgrokProcess().isRunning());
+        ngrokClientV2.getNgrokProcess().setAuthToken(ngrokAuthToken);
+        assertFalse(ngrokClientV2.getNgrokProcess().isRunning());
         final CreateTunnel createTunnel = new CreateTunnel.Builder()
                 .withAddr("file:///")
                 .build();
-        final String publicUrl = ngrokClientV3.connect(createTunnel).getPublicUrl();
+        final String publicUrl = ngrokClientV2.connect(createTunnel).getPublicUrl();
         Thread.sleep(1000);
 
         // WHEN
-        ngrokClientV3.disconnect(publicUrl);
+        ngrokClientV2.disconnect(publicUrl);
         Thread.sleep(1000);
-        final List<Tunnel> tunnels = ngrokClientV3.getTunnels();
+        final List<Tunnel> tunnels = ngrokClientV2.getTunnels();
 
         // THEN
-        assertTrue(ngrokClientV3.getNgrokProcess().isRunning());
+        assertTrue(ngrokClientV2.getNgrokProcess().isRunning());
         // There is still one tunnel left, as we only disconnected the http tunnel
         assertEquals(1, tunnels.size());
     }
@@ -756,6 +758,80 @@ class NgrokClientTest extends NgrokTestCase {
     }
 
     @Test
+    public void testTunnelDefinitionsV3CloudEdge() {
+        final String ngrokAuthToken = System.getenv("NGROK_AUTHTOKEN");
+        assumeTrue(isNotBlank(System.getenv("NGROK_AUTHTOKEN")), "NGROK_AUTHTOKEN environment variable not set");
+        final String ngrokHttpEdge = System.getenv("NGROK_HTTP_EDGE");
+        assumeTrue(isNotBlank(System.getenv("NGROK_HTTP_EDGE")), "NGROK_HTTP_EDGE environment variable not set");
+        final String ngrokTcpEdge = System.getenv("NGROK_TCP_EDGE");
+        assumeTrue(isNotBlank(System.getenv("NGROK_TCP_EDGE")), "NGROK_TCP_EDGE environment variable not set");
+        final String ngrokApiKey = System.getenv("NGROK_API_KEY");
+        assumeTrue(isNotBlank(System.getenv("NGROK_API_KEY")), "NGROK_API_KEY environment variable not set");
+        final String ngrokHttpEdgeEndpoint = System.getenv("NGROK_HTTP_EDGE_ENDPOINT");
+        assumeTrue(isNotBlank(System.getenv("NGROK_HTTP_EDGE_ENDPOINT")), "NGROK_HTTP_EDGE_ENDPOINT environment variable not set");
+        final String ngrokTcpEdgeEndpoint = System.getenv("NGROK_TCP_EDGE_ENDPOINT");
+        assumeTrue(isNotBlank(System.getenv("NGROK_TCP_EDGE_ENDPOINT")), "NGROK_TCP_EDGE_ENDPOINT environment variable not set");
+
+        // GIVEN
+        final Map<String, Object> edgeHttpTunnelConfig = Map.of(
+                "addr", "80",
+                "labels", List.of(String.format("edge=%s", ngrokHttpEdge)));
+        final Map<String, Object> edgeTcpTunnelConfig = Map.of(
+                "addr", "22",
+                "labels", List.of(String.format("edge=%s", ngrokTcpEdge)));
+        final Map<String, Object> tunnelsConfig = Map.of(
+                "edge-http-tunnel", edgeHttpTunnelConfig,
+                "edge-tcp-tunnel", edgeTcpTunnelConfig);
+        final Map<String, Object> config = Map.of("tunnels", tunnelsConfig);
+
+        final Path configPath2 = Paths.get(javaNgrokConfigV3.getConfigPath().getParent().toString(), "config2.yml");
+        ngrokInstaller.installDefaultConfig(configPath2, config, javaNgrokConfigV3.getNgrokVersion());
+        final JavaNgrokConfig javaNgrokConfig2 = new JavaNgrokConfig.Builder(javaNgrokConfigV3)
+                .withConfigPath(configPath2)
+                .withAuthToken(ngrokAuthToken)
+                .withApiKey(ngrokApiKey)
+                .build();
+        ngrokProcessV3_2 = new NgrokProcess(javaNgrokConfig2, ngrokInstaller);
+        final NgrokClient ngrokClient2 = new NgrokClient.Builder()
+                .withJavaNgrokConfig(javaNgrokConfig2)
+                .withNgrokProcess(ngrokProcessV3_2)
+                .build();
+
+        // WHEN
+        final CreateTunnel createHttpEdgeTunnel = new CreateTunnel.Builder()
+                .withNgrokVersion(NgrokVersion.V3)
+                .withName("edge-http-tunnel")
+                .build();
+        final CreateTunnel createTcpEdgeTunnel = new CreateTunnel.Builder()
+                .withNgrokVersion(NgrokVersion.V3)
+                .withName("edge-tcp-tunnel")
+                .build();
+        final Tunnel httpEdgeTunnel = ngrokClient2.connect(createHttpEdgeTunnel);
+        final Tunnel tcpEdgeTunnel = ngrokClient2.connect(createTcpEdgeTunnel);
+        final List<Tunnel> tunnels = ngrokClient2.getTunnels();
+        tunnels.sort(Comparator.comparing(Tunnel::getProto));
+
+        // THEN
+        assertEquals("edge-http-tunnel", httpEdgeTunnel.getName());
+        assertEquals("http://localhost:80", httpEdgeTunnel.getConfig().getAddr());
+        assertEquals("https", httpEdgeTunnel.getProto());
+        assertEquals(ngrokHttpEdgeEndpoint, httpEdgeTunnel.getPublicUrl());
+        assertEquals("edge-tcp-tunnel", tcpEdgeTunnel.getName());
+        assertEquals("tcp://localhost:22", tcpEdgeTunnel.getConfig().getAddr());
+        assertEquals("tcp", tcpEdgeTunnel.getProto());
+        assertEquals(ngrokTcpEdgeEndpoint, tcpEdgeTunnel.getPublicUrl());
+        assertEquals(2, tunnels.size());
+        assertEquals("edge-http-tunnel", tunnels.get(0).getName());
+        assertEquals("http://localhost:80", tunnels.get(0).getConfig().getAddr());
+        assertEquals("https", tunnels.get(0).getProto());
+        assertEquals(ngrokHttpEdgeEndpoint, tunnels.get(0).getPublicUrl());
+        assertEquals("edge-tcp-tunnel", tunnels.get(1).getName());
+        assertEquals("tcp://localhost:22", tunnels.get(1).getConfig().getAddr());
+        assertEquals("tcp", tunnels.get(1).getProto());
+        assertEquals(ngrokTcpEdgeEndpoint, tunnels.get(1).getPublicUrl());
+    }
+
+    @Test
     public void testTunnelDefinitionsV3OAuth() {
         final String ngrokAuthToken = System.getenv("NGROK_AUTHTOKEN");
         assumeTrue(isNotBlank(System.getenv("NGROK_AUTHTOKEN")), "NGROK_AUTHTOKEN environment variable not set");
@@ -838,14 +914,14 @@ class NgrokClientTest extends NgrokTestCase {
         final Tunnel ngrokTunnel2 = ngrokClient2.connect(createTunnelSubdomain);
 
         // THEN
-        assertEquals("java-ngrok-default (http)", ngrokTunnel1.getName());
+        assertEquals("java-ngrok-default", ngrokTunnel1.getName());
         assertEquals("http://localhost:8080", ngrokTunnel1.getConfig().getAddr());
-        assertEquals("http", ngrokTunnel1.getProto());
-        assertEquals(String.format("http://%s.ngrok.io", subdomain1), ngrokTunnel1.getPublicUrl());
-        assertEquals("java-ngrok-default (http)", ngrokTunnel2.getName());
+        assertEquals("https", ngrokTunnel1.getProto());
+        assertEquals(String.format("https://%s.ngrok.io", subdomain1), ngrokTunnel1.getPublicUrl());
+        assertEquals("java-ngrok-default", ngrokTunnel2.getName());
         assertEquals("http://localhost:5000", ngrokTunnel2.getConfig().getAddr());
-        assertEquals("http", ngrokTunnel2.getProto());
-        assertEquals(String.format("http://%s.ngrok.io", subdomain2), ngrokTunnel2.getPublicUrl());
+        assertEquals("https", ngrokTunnel2.getProto());
+        assertEquals(String.format("https://%s.ngrok.io", subdomain2), ngrokTunnel2.getPublicUrl());
     }
 
     @Test